requirements.txt
annotated-types==0.7.0
ansible-base==2.10.8
anyio==4.4.0
apache-libcloud==3.2.0
argcomplete==1.8.1
awscli==1.22.34
Babel==2.8.0
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
docutils==0.17.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.6.1
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
ntlm-auth==1.4.0
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
orjson==3.10.6
packaging==21.3
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.11.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.1
pywinrm==0.3.0
PyYAML==5.4.1
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rich==13.7.1
roman==3.3
rsa==4.8
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
selinux==3.3
shellingham==1.5.4
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
systemd-python==234
threadpoolctl==3.5.0
torch==2.3.1
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
whisper==1.1.10
xmltodict==0.12.0
zipp==1.0.0
